refactor: split launcher and reporter into separate binaries
