+++ Cargo.toml
@@ -3,6 +3,15 @@ name = "app-activate"
 version = "0.5.0"
 edition = "2021"
 description = "A minimal application launcher, just for my needs."
+default-run = "app-activate"
+
+[[bin]]
+name = "app-activate"
+path = "src/main.rs"
+
+[[bin]]
+name = "app-activate-reporter"
+path = "src/reporter.rs"
 
 [dependencies]
 # Error handling
