+++ src/args.rs
@@ -24,7 +24,4 @@ pub enum Command {
 
     /// Unregister the application from starting on login.
     Unregister,
-
-    /// Report launch history if available.
-    Report,
 }
